+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/controller/admin/school/AdminSchoolController.java
@@ -4,7 +4,9 @@ import ProjectDoge.StudentSoup.dto.school.SchoolFormDto;
 import ProjectDoge.StudentSoup.dto.school.SchoolSearch;
 import ProjectDoge.StudentSoup.entity.school.School;
 import ProjectDoge.StudentSoup.repository.school.SchoolRepository;
-import ProjectDoge.StudentSoup.service.SchoolService;
+import ProjectDoge.StudentSoup.service.admin.AdminSchoolService;
+import ProjectDoge.StudentSoup.service.school.SchoolFindService;
+import ProjectDoge.StudentSoup.service.school.SchoolRegisterService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -20,9 +22,9 @@ import java.util.List;
 @Controller
 @RequiredArgsConstructor
 public class AdminSchoolController {
-
-    private final SchoolService schoolService;
-
+    private final SchoolFindService schoolFindService;
+    private final SchoolRegisterService schoolRegisterService;
+    private final AdminSchoolService adminSchoolService;
     private final SchoolRepository schoolRepository;
 
     @GetMapping("admin/school/new")
@@ -35,29 +37,29 @@ public class AdminSchoolController {
         if(bindingResult.hasErrors()){
             return "/admin/school/createSchool";
         }
-        schoolService.join(formDto);
+        schoolRegisterService.join(formDto);
         return "redirect:/admin/schools";
     }
     @GetMapping("/admin/schools")
     public String schoolList(@ModelAttribute("schoolSearch") SchoolSearch schoolSearch, Model model) {
-        List<School> schools = schoolService.findAll();
+        List<School> schools = schoolFindService.findAll();
         model.addAttribute("schools", schools);
 
-        List<School> findSchools = schoolService.AdminSearchSchools(schoolSearch);
+        List<School> findSchools = adminSchoolService.AdminSearchSchools(schoolSearch);
         model.addAttribute("findSchools", findSchools);
 
         return "/admin/school/schoolList";
     }
     @GetMapping("/admin/school/edit")
     public String editSchool(@RequestParam("schoolId")Long schoolId,Model model){
-        SchoolFormDto updateSchool = schoolService.AdminFindUpdateSchool(schoolId);
+        SchoolFormDto updateSchool = adminSchoolService.AdminFindUpdateSchool(schoolId);
         model.addAttribute("schoolId",schoolId);
         model.addAttribute("schoolForm",updateSchool);
         return "/admin/school/updateSchool";
     }
     @PostMapping("/admin/school/edit")
     public String editSchool(@RequestParam("schoolId")Long schoolId,@ModelAttribute("form") SchoolFormDto schoolFormDto){
-        schoolService.AdminUpdateSchool(schoolId,schoolFormDto);
+        adminSchoolService.AdminUpdateSchool(schoolId,schoolFormDto);
         return "redirect:/admin/schools";
     }
     @GetMapping("/admin/school/delete")
