+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/controller/member/MemberUpdateController.java
@@ -5,7 +5,8 @@ import ProjectDoge.StudentSoup.dto.member.MemberUpdateDto;
 import ProjectDoge.StudentSoup.dto.member.MemberUpdateValidationDto;
 import ProjectDoge.StudentSoup.entity.member.Member;
 import ProjectDoge.StudentSoup.service.member.MemberFindService;
-import ProjectDoge.StudentSoup.service.member.MemberService;
+import ProjectDoge.StudentSoup.service.member.MemberUpdateService;
+import ProjectDoge.StudentSoup.service.member.MemberValidationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.BindingResult;
@@ -17,16 +18,16 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 @RequestMapping("/members")
 public class MemberUpdateController {
-
-    private final MemberService memberService;
+    private final MemberValidationService memberValidationService;
     private final MemberFindService memberFindService;
+    private final MemberUpdateService memberUpdateService;
 
     @PostMapping("/edit/{memberId}/validation")
     public String editIdCheck(@PathVariable("memberId") Long memberId,
                               @Validated @RequestBody MemberUpdateValidationDto dto,
                               BindingResult bindingResult){
         Member member = memberFindService.findOne(dto.getMemberId());
-        memberService.validationCoincideMemberIdPwd(member, dto.getPwd());
+        memberValidationService.validationCoincideMemberIdPwd(member, dto.getPwd());
 
         return "ok";
     }
@@ -34,7 +35,7 @@ public class MemberUpdateController {
     @PostMapping("/edit/{memberId}")
     public MemberDto editMember(@PathVariable("memberId") Long memberId,
                                 @RequestBody MemberUpdateDto dto){
-        Long updatedMemberId = memberService.updateMember(dto);
+        Long updatedMemberId = memberUpdateService.updateMember(dto);
         Member member = memberFindService.findOne(updatedMemberId);
         MemberDto memberDto = new MemberDto().getMemberDto(member);
         return memberDto;
