+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/controller/member/ProfileController.java
@@ -1,7 +1,7 @@
 package ProjectDoge.StudentSoup.controller.member;
 
 import ProjectDoge.StudentSoup.dto.member.MemberDto;
-import ProjectDoge.StudentSoup.service.member.MemberService;
+import ProjectDoge.StudentSoup.service.member.MemberProfileImageUpdateService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
@@ -13,12 +13,12 @@ import org.springframework.web.multipart.MultipartFile;
 @RequiredArgsConstructor
 public class ProfileController {
 
-    private final MemberService memberService;
+    private final MemberProfileImageUpdateService memberProfileImageUpdateService;
 
     @PostMapping(value = "/members/edit/image", consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE})
     public MemberDto updateProfile(@RequestParam(value = "memberId") Long memberId, @RequestPart MultipartFile multipartFile){
         log.info("받은 memberId : [{}], imageFile : [{}]", memberId, multipartFile.getOriginalFilename());
-
-        return memberService.memberProfileUpdate(memberId, multipartFile);
+        MemberDto dto = memberProfileImageUpdateService.memberProfileUpdate(memberId, multipartFile);
+        return dto;
     }
 }
