+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/service/admin/AdminMemberService.java
@@ -1,14 +1,17 @@
 package ProjectDoge.StudentSoup.service.admin;
 
 import ProjectDoge.StudentSoup.dto.admin.AdminMemberUpdateForm;
-import ProjectDoge.StudentSoup.dto.member.MemberFormBDto;
 import ProjectDoge.StudentSoup.dto.member.MemberUpdateDto;
+import ProjectDoge.StudentSoup.entity.file.ImageFile;
 import ProjectDoge.StudentSoup.entity.member.Member;
 import ProjectDoge.StudentSoup.entity.school.Department;
 import ProjectDoge.StudentSoup.entity.school.School;
 import ProjectDoge.StudentSoup.exception.member.MemberNotFoundException;
-import ProjectDoge.StudentSoup.exception.member.MemberValidationException;
 import ProjectDoge.StudentSoup.repository.member.MemberRepository;
+import ProjectDoge.StudentSoup.service.file.FileService;
+import ProjectDoge.StudentSoup.service.department.DepartmentFindService;
+import ProjectDoge.StudentSoup.service.member.MemberValidationService;
+import ProjectDoge.StudentSoup.service.school.SchoolFindService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -22,6 +25,10 @@ import javax.transaction.Transactional;
 public class AdminMemberService {
 
     private final MemberRepository memberRepository;
+    private final MemberValidationService memberValidationService;
+    private final SchoolFindService schoolFindService;
+    private final DepartmentFindService departmentFindService;
+    private final FileService fileService;
 
     @Transactional
     public Long adminMemberUpdate(AdminMemberUpdateForm dto, MultipartFile file) {
@@ -32,8 +39,8 @@ public class AdminMemberService {
                     throw new MemberNotFoundException();
                 });
 
-        School school = getMemberDtoSchool(dto.getSchoolId());
-        Department department = getMemberDtoDepartment(dto.getDepartmentId());
+        School school = schoolFindService.findOne(dto.getSchoolId());
+        Department department = departmentFindService.findOne(dto.getDepartmentId());
 
         log.info("현재 가져온 학교와 학과 : [{}] / [{}]", school.getSchoolName(), department.getDepartmentName());
         validationChangedNicknameEmail(dto, member);
@@ -44,6 +51,17 @@ public class AdminMemberService {
 
         return member.getMemberId();
     }
+    public void validationChangedNicknameEmail(MemberUpdateDto dto, Member member) {
+        log.info("회원 업데이트 중 닉네임과 이메일 검증을 시작합니다.");
+        if(!member.getNickname().equals(dto.getNickname())) {
+            memberValidationService.validateDuplicateMemberNickname(dto.getNickname());
+        }
+        if(!member.getEmail().equals(dto.getEmail())) {
+            memberValidationService.validateDuplicateMemberEmail(dto.getEmail());
+        }
+        log.info("회원 업데이트 중 닉네임 이메일 검증이 완료되었습니다.");
+    }
+
 
     private void adminUpdateMemberField(MemberUpdateDto dto, Member member) {
         member.setPwd(dto.getPwd());
@@ -51,36 +69,10 @@ public class AdminMemberService {
         member.setNickname(dto.getNickname());
     }
 
-    private void validationChangedNicknameEmail(MemberUpdateDto dto, Member member) {
-        log.info("회원 업데이트 중 닉네임과 이메일 검증을 시작합니다.");
-        if(!member.getNickname().equals(dto.getNickname())) {
-            validateDuplicateMemberNickname(dto.getNickname());
+    private void updateMemberProfileImage(Member member, Long fileId) {
+        if(fileId != null) {
+            ImageFile imageFile = fileService.findOne(fileId);
+            member.setImageFile(imageFile);
         }
-        if(!member.getEmail().equals(dto.getEmail())) {
-            validateDuplicateMemberEmail(dto.getEmail());
-        }
-        log.info("회원 업데이트 중 닉네임 이메일 검증이 완료되었습니다.");
-    }
-
-    private void validateDuplicateMemberNickname(String nickname) {
-        log.info("회원 닉네임 중복 검증 메소드가 실행되었습니다.");
-
-        Member findMember = memberRepository.findByNickname(nickname);
-        if (findMember != null) {
-            log.info("회원 닉네임이 중복 예외가 발생했습니다.");
-            throw new MemberValidationException("중복된 닉네임입니다.");
-        }
-        log.info("회원 닉네임 중복 검증이 완료되었습니다.");
-    }
-
-    private void validateDuplicateMemberEmail(String email) {
-        log.info("회원 이메일 중복 검증 메소드가 실행되었습니다.");
-
-        Member findMember = memberRepository.findByEmail(email);
-        if (findMember != null) {
-            log.info("회원 이메일 중복 예외가 발생했습니다.");
-            throw new MemberValidationException("중복된 이메일입니다.");
-        }
-        log.info("회원 이메일 중복 검증이 완료되었습니다.");
     }
 }
