+++ StudentSoup/src/test/java/ProjectDoge/StudentSoup/member/MemberUpdateTest.java
@@ -1,0 +1,15 @@
+package ProjectDoge.StudentSoup.member;
+
+import lombok.extern.slf4j.Slf4j;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import javax.transaction.Transactional;
+
+@Slf4j
+@SpringBootTest
+@Transactional
+public class MemberUpdateTest {
+
+
+}
